+++ src/DI/Definition/Source/ArrayDefinitionSource.php
@@ -12,7 +12,6 @@ namespace DI\Definition\Source;
 use DI\Definition\ArrayDefinition;
 use DI\Definition\ClassDefinition;
 use DI\Definition\Definition;
-use DI\Definition\MergeableDefinition;
 use DI\Definition\ValueDefinition;
 use DI\Definition\Helper\DefinitionHelper;
 
@@ -21,7 +20,7 @@ use DI\Definition\Helper\DefinitionHelper;
  *
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
  */
-class ArrayDefinitionSource implements ChainableDefinitionSource
+class ArrayDefinitionSource extends ChainableDefinitionSource implements DefinitionSource
 {
     const WILDCARD = '*';
     /**
@@ -30,41 +29,10 @@ class ArrayDefinitionSource implements ChainableDefinitionSource
     const WILDCARD_PATTERN = '([^\\\\]+)';
 
     /**
-     * @var DefinitionSource
-     */
-    private $chainedSource;
-
-    /**
      * DI definitions in a PHP array
      * @var array
      */
     private $definitions = array();
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDefinition($name)
-    {
-        $definition = $this->findDefinition($name);
-
-        if ($definition === null) {
-            // Not found, we use the chain or return null
-            if ($this->chainedSource) {
-                return $this->chainedSource->getDefinition($name);
-            }
-            return null;
-        }
-
-        // Enrich definition in sub-source
-        if ($this->chainedSource && $definition instanceof MergeableDefinition) {
-            $subDefinition = $this->chainedSource->getDefinition($definition->getExtendedDefinitionName());
-            if ($subDefinition instanceof MergeableDefinition) {
-                $definition = $definition->merge($subDefinition);
-            }
-        }
-
-        return $definition;
-    }
 
     /**
      * @param array $definitions DI definitions in a PHP array indexed by the definition name.
@@ -85,18 +53,10 @@ class ArrayDefinitionSource implements ChainableDefinitionSource
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function chain(DefinitionSource $source)
-    {
-        $this->chainedSource = $source;
-    }
-
-    /**
      * @param string $name
      * @return Definition|null
      */
-    private function findDefinition($name)
+    protected function findDefinition($name)
     {
         // Look for the definition by name
         if (array_key_exists($name, $this->definitions)) {