+++ change-log.md
@@ -11,6 +11,10 @@ BC breaks:
 
 - [#198](https://github.com/mnapoli/PHP-DI/issues/198) `ocramius/proxy-manager` is not installed by default anymore, you need to require it in `composer.json` if you want to use **lazy injection**
 
+Internal changes in case you were replacing/extending some parts:
+
+- `DI\Definition\Source\ChainableDefinitionSource` is now an abstract class (was previously an interface).
+
 ## 4.4
 
 Read the [news entry](news/13-php-di-4-4-released.md).